Finish button class.
NOTE: The class really doesn't have anything to do with buttons. A different name might be more suitable.

The button class is used for a button that contains an image and text.

There are four different kinds of buttons:
* button-left-aligned
* button-right-aligned
* button-top-center-aligned
* button-bottom-center-aligned

The direction specified in the class name determines where the image is relative to the text.

The left-aligned button is justified to the left of its containing box.
The right-aligned button is justified to the right of its containing box.

The x-center-aligned buttons are center-justified in its containing box.